requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
requests
langchain
langchain_community
ollama 
langchain_openai




